perf(memory): float32 not float64 for full-image buffers (tiled, spcc)
- tiled.py (AI inference output/weight accumulators) and spcc.py (color-cal
  result buffer) allocated full-resolution float64, then cast to float32 at the
  end. [0,1] image accumulation needs no 64-bit precision. float32 halves these
  full-frame buffers (~280MB each mono in tiled; ~440MB RGB in spcc at 70MP).
71 spcc/color-cal/super-res/sharpen tests pass.

+++ astraios/ai/inference/tiled.py
@@ -80,9 +80,10 @@ def tiled_inference(
     n_tiles = len(tiles)
     log.debug("Tiled inference: %d tiles of %dx%d with %dpx overlap", n_tiles, tile_size, tile_size, overlap)
 
-    # Output and weight buffers
-    output = np.zeros((h, w), dtype=np.float64)
-    weight = np.zeros((h, w), dtype=np.float64)
+    # Output and weight buffers — float32 is plenty for [0,1] tile accumulation
+    # (the result is cast to float32 anyway) and halves these full-image buffers.
+    output = np.zeros((h, w), dtype=np.float32)
+    weight = np.zeros((h, w), dtype=np.float32)
 
     # Blending weight: cosine ramp at edges
     blend = _create_blend_weight(tile_size, overlap)
